requirements.txt
keyboard==0.13.5
matplotlib==3.3.1
numpy==1.19.1
opencv-python==4.4.0.42
Pillow==7.2.0
pyparsing==2.4.7
pyserial==3.4
python-dateutil==2.8.1
